refactored - add new folder structure, change type path, refactor some files, add barrels
