+++ src/components/Header/DarkModeToggle/index.ts
@@ -1,0 +1,1 @@
+export { default} from './DarkModeToggle'
