+++ src/components/Job/JobCard/JobCard.styled.tsx
