+++ src/components/Job/JobView/JobView.styled.tsx
