+++ src/components/Job/JobView/index.ts
@@ -1,0 +1,1 @@
+export { default } from './JobView';
