+++ src/components/Pages/MainPage.tsx
@@ -1,6 +1,6 @@
 
 import Header from '../Header/Header';
-import JobList from '../JobList/JobList';
+import JobList from '../Job/JobList';
 
 function MainPage() {
 	return (
