+++ src/components/UI/Button/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Button';
