+++ src/components/UI/CheckBox/Checkbox.tsx
@@ -1,12 +1,12 @@
-import CheckIcon from './icon/CheckIcon';
 import { StyledCheckbox, StyledCheckmark } from './Checkbox.style';
-import { CheckboxType } from '../../../types';
+import { CheckboxProps } from './Checkbox.type';
+import CheckIcon from './icon/CheckIcon';
 
-const Checkbox = ({ name, id, onClick, checked }: CheckboxType) => (
-	<StyledCheckmark checked={checked} onClick={onClick}>
-		<StyledCheckbox type='checkbox' name={name} id={id} />
-		{checked && <CheckIcon />}
-	</StyledCheckmark>
+const Checkbox = ({ name, id, onClick, checked }: CheckboxProps) => (
+  <StyledCheckmark checked={checked} onClick={onClick}>
+    <StyledCheckbox type="checkbox" name={name} id={id} />
+    {checked && <CheckIcon />}
+  </StyledCheckmark>
 );
 
 export default Checkbox;
