+++ src/components/UI/Modal/Modal.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { ModalType } from '../../../types';
 import { StyledModal, StyledOverlay } from './Modal.styled';
+import { ModalProps } from './Modal.type';
 
-export default function Modal({ backdropHandler, children }: ModalType) {
+export default function Modal({ backdropHandler, children }: ModalProps) {
 	return ReactDOM.createPortal(
 		<>
 			<StyledOverlay
-				onClick={() => {
-					backdropHandler(false);
-				}}
+				onClick={() => backdropHandler(false)}
 			></StyledOverlay>
 			<StyledModal>{children}</StyledModal>
 		</>,
