+++ src/components/UI/Modal/Modal.type.ts
@@ -1,0 +1,4 @@
+export interface ModalProps {
+  backdropHandler: React.Dispatch<React.SetStateAction<boolean>>;
+  children: JSX.Element;
+}
