+++ src/components/UI/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from './Button';
+export { default as Checkbox } from './Checkbox';
+export { default as Input } from './Input';
+export { default as Modal } from './Modal';
