--- src/types/Modal.type.ts
@@ -1,4 +1,0 @@
-export interface ModalType {
-    backdropHandler: React.Dispatch<React.SetStateAction<boolean>>;
-    children: JSX.Element;
-  }
