--- src/types/index.ts
@@ -1,5 +1,0 @@
-export * from './Checkbox.type';
-export * from './JobCard.type';
-export * from './ButtonProps.type';
-export * from './Input.type';
-export * from './Modal.type';
